25/07/2020

+++ src/components/Pdv.vue
@@ -32,21 +32,24 @@
         </li>
       </ul>
 </div>
+<!-- Layoute para modelo simples -->
  <div  v-if="simples" class="container">
         <input type="text" class="input1" v-model="t1" placeholder="Coloque um número para a mesa">
-        <button name="button" @click="adicionar_identificador2(t1)"class="material-icons botao">navigate_next </button>
+        <button name="button" @click="executar_simples(operacao,t1)"class="material-icons botao">navigate_next </button>
     </div>
-       <div v-if="!simples" class="container">
+
+    <!-- Layoute para modelo Duplo -->
+       <div v-if="duplo" class="container">
         <input type="text" class="input1" v-model="t1" placeholder="Coloque a primeira mesa">
         <h4>{{operacao}} com :</h4>
           <input type="text" class="input1" v-model="t2" placeholder="Coloque a segunda mesa ">
-        <button name="button" @click="juntar_identificador(t1,t2)"class="material-icons botao">navigate_next </button>
+        <button name="button" @click="executar_duplo(operacao,t1,t2)"class="material-icons botao">navigate_next </button>
       </div>
 
     <h1>Abertas</h1>
     <ul  v-for="(atributo,index) in mesa">
       <li>
-        <div class="card" @dblclick="comanda(atributo.identificador)" >
+        <div class="botao_bola" @dblclick="comanda(atributo.identificador)" >
             <h4 >{{atributo.identificador}}</h4>
           </div>
       </li>
@@ -74,6 +77,7 @@ return{
   mesa:[],
   t3:'1',
   simples:null,
+  duplo:null,
   operacao:null,
 }
 },
@@ -81,71 +85,82 @@ methods:{
 
 adicionar(){
   this.simples = true,
+  this.duplo = false,
   this.operacao = 'adicionar'
 },
 excluir(){
   this.simples = true,
+  this.duplo = false,
   this.operacao = 'excluir'
 },
 trocar(){
   this.simples = false,
+  this.duplo = true,
   this.operacao = 'trocar'
 },
 juntar(){
   this.simples = false,
+  this.duplo = true,
   this.operacao = 'juntar'
 },
+// executar_simples(operacao){
+// switch (operacao) {
+//   case adicionar:
+//   alert('sendo executado')
+//     break;
+//     case excluir:
+//     alert('sendo executado')
+//
+//     // code block
+//     break;
+//   default:
+//   alert('deu erro')
+//
+// }
+// },
 
+//executando modelo simples
+executar_simples(operacao, t1){
+if (operacao == 'adicionar') {
 
-  //Criando metado adicionar_identificador
-
-  adicionar_identificador(t){
-  db.collection('mesa').add({
-    identificador: this.t,
+  db.collection('mesa').doc(this.t1).set({
+    identificador: this.t1,
   }).then(()=>{
     alert('Mesa criada com sucesso')
     location.reload();
   }
-).catch(err=>{
+  ).catch(err=>{
   console.log(err)
-})
+  })
+
+  } else if (operacao == 'excluir') {
+    // selecionando o documento e excluindo
+    //---------------------------- aperfeiçoar, para não ter q fazer reload-----------------------------------------
+        let excluir = db.collection('mesa').doc(this.t1).delete().then(()=>{
+          location.reload();
+        // this.mesa = this.mesa.filter(curso=>{
+        // return curso.id != id
+      }).catch(err=>
+      console.log(err))
+
+
+}else{
+alert('deu errado')}
 },
 
-adicionar_identificador2(t1){
-db.collection('mesa').doc(this.t1).set({
-  identificador: this.t1,
-}).then(()=>{
-  alert('Mesa criada com sucesso')
-  location.reload();
-}
-).catch(err=>{
-console.log(err)
-})
+  //executando modelo duplo
+executar_duplo(operacao,t1,t2){
+if (operacao == 'juntar') {
+  db.collection('mesa').doc(this.t2).set({
+    identificador: this.t1,
+  });
+  alert('Mesa '+ this.t1 +' trocada por '+ this.t2)
+  }
+
 },
 comanda(identificador){
   this.$router.push({ name: 'Mesa', params:{mesa:identificador}})
 },
-
-// criando metado juntar_identificador
-
-juntar_identificador(t1,t2){
-
-db.collection('mesa').doc(this.t2).set({
-  identificador: this.t1,
-});
-alert('Mesa '+ this.t1 +' trocada por '+ this.t2)
-},
-excluir_identificador2(t1){
-// selecionando o documento e excluindo
-//---------------------------- aperfeiçoar, para não ter q fazer reload-----------------------------------------
-    let excluir = db.collection('mesa').doc(this.t1).delete().then(()=>{
-      location.reload();
-    // this.mesa = this.mesa.filter(curso=>{
-    // return curso.id != id
-  }).catch(err=>
-  console.log(err))
-
-  }
   },
   // recuperando dados do firebase
 created(){
@@ -164,24 +179,26 @@ created(){
 
 </script>
 
+
+
 <style scoped>
-.card {
+.botao_bola {
   height: 50px;
   width: 50px;
   border-radius: 50%;
   background-color: #009688;
-  /* Add shadows to create the "card" effect */
+  /* Add shadows to create the "botao_bola" effect */
   box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
   transition: 0.3s;
 }
 
 /* On mouse-over, add a deeper shadow */
-.card:hover {
+.botao_bola:hover {
   box-shadow: 0 8px 16px 0 #000000;
 }
 
 
-/* Add some padding inside the card container */
+/* Add some padding inside the botao_bola container */
 
 .material-icons.md-40-light {
   text-align: center;
